+++ src/docker/mod.rs
@@ -11,3 +11,4 @@ pub mod setup;
 pub mod state_machine_handlers;
 pub mod stop_app;
 pub mod ttl_checker;
+pub mod validation;
